requirements.txt
geopandas==0.14.4
numpy==1.26.4
requests==2.32.3
xarray==2024.5.0
